Finished writing and testing the generic parser for bro logs.
Wrote core functions ParseFields() and BufferRow().
The combination of these functions allow this package
to parse through bro logs and access the rows by ranging
over parser.Row(). An example of this can be found in
TestBufferRow().

Added godocs for each function.

Added a sample test log to parse through.