Started setting up db testing with docker containers
Added functions InitDB() and ConnectMySql() to db package.
ConnectMySql() keeps trying to initiate a connection
to the db until it succeeds. Added config files for db testing, and a
sample schema. Added a docker-compose file which spins up
an instance of mysql. Added vendoring to project. Added
test.py file that controls testing the db package with docker
containers.

This is in progress and not yet completed