Added benchmarks to the parser and tests package
Added a function to automatically create a buffer size
equivalent to the number of lines in a log file to be parsed.
Can now run benchmarks via test.py
Added config files to specific location of log files to test and
benchmark in parser/ and tests/
Updated readme to include explanation of tests and benchmarks.